that the connections are correctly terminated.
We use OpenSSL for the tests and modern OpenSSL versions don't
support SSLv{1,2,3}.0, so use TLSv1.{0,1} just to test that we
correctly drop wrong TLS connections. We do not support SSL as well
and any SSL record is treated as a broken TLS record, so fuzzing of
normal TLS fields should be used to test TLS fields processing.