Fix incorrect find usage.
`not a.find(b)` is used incorrectly: `find()` returns index of element,
where the `b` was found in `a`, and `-1` if not found. `listen` directive
can't have a value starting from `https`, so the whole condition is always
false and `KeyError` exception is thrown on accessing absent
`tls_certificate` directive.