Test MatchHost has been rewriten
The previous approach that used chains replaced with no chains, because chains
is a semi-depricated part that belongs to old framework. Full request
comparision dropped, because request can be changed by tempesta during
forwarding thus full comparision is not a valid approach. Now just assign id
for each request and check it on receive, it's enough for current test.
The goal of test is verify correctnes of HTTPTables rules and routing
without checking correctnes of forwarded message.