Enabled and fix tests blocked by #1762
`client_header_timeout` and `client_body_timeout`
configuration options work in the following way:
When we receive the first block of the request
header or body, we remember the time when it was
received. When the next block arrives, we check
the time of its arrival and if the waiting timeout
exceeds the specified one, we drop the request.
Fix tests according this description.