Workaround for WS pipelined test
Implemented workaround to avoid blinking of WS pipeline
tests on remote and segmantation setup. Now we just not
parse the response, because right after the response
might follow data related to WS protocol. In this test
we might ensure that connection is closed and there is
a warning in dmesg.

Enabled WS pipelined test for tcp segmentation. On tcp
segmantation this test covers another branch of the
code, that not covered in regular tests. Rest WS tests
are disabled.