2466: Add ability to specify clickhouse http port and tcp port using config
The 'clickhouse_port' configuration property was intended for use by both the
test and the tfw_logger. However, this is not possible because they use
different protocols. The tests use the Clickhouse HTTP API, while the logger
uses the internal API, also referred to as the TCP API. To differentiate
between these two approaches, the "clickhouse_port" configuration property was
replaced with two new properties: "clickhouse_http_port" and
"clickhouse_tcp_port."